Organize use cases in test/support
