+++ test/micro/case/flow/reducer_test.rb
@@ -1,188 +1,7 @@
 require 'test_helper'
+require 'support/todoing/boot'
 
 class Micro::Case::Flow::ReducerTest < Minitest::Test
-  require 'digest'
-  require 'securerandom'
-
-  class Base
-    attr_reader :id
-
-    def self.__relation; (@relation ||= []); end
-
-    def self.all; __relation.to_a; end
-
-    def self.count; __relation.size; end
-
-    def self.delete_all; @relation = []; end
-
-    def self.find_by_id(id)
-      __relation.find { |rec| rec.id == id }
-    end
-
-    def new_record?; id.nil?; end
-
-    private
-
-      def save_new_record
-        if new_record?
-          @id = SecureRandom.uuid
-
-          yield
-
-          self.class.__relation << self
-        end
-
-        true
-      end
-  end
-
-  class User < Base
-    attr_reader :password_hash
-    attr_accessor :name
-
-    def initialize(options = {})
-      @name = options[:name]
-      @password = options[:password]
-    end
-
-    def invalid?
-      name.empty? || @password.empty?
-    end
-
-    def save
-      return false if invalid?
-
-      self.name = name
-
-      save_new_record do
-        @password_hash = Digest::SHA256.hexdigest(@password)
-      end
-    end
-
-    def wrong_password?(value)
-      password_hash != Digest::SHA256.hexdigest(value)
-    end
-  end
-
-  class Todo < Base
-    attr_accessor :description, :done, :user_id
-
-    def self.find_by_id_and_user_id(id, user_id)
-      all.find { |todo| todo.id == id && todo.user_id && user_id }
-    end
-
-    def initialize(options = {})
-      @user_id = options[:user_id]
-      @description = options[:description]
-    end
-
-    def invalid?
-      description.empty? || user_id.empty?
-    end
-
-    def save
-      return false if invalid?
-
-      self.description = description
-
-      save_new_record { @done = done? }
-    end
-
-    def pending?; !done; end
-
-    def done?; !pending?; end
-  end
-
-  module Users
-    class Create < Micro::Case::Strict
-      attributes :name, :password, :password_confirmation
-
-      def call!
-        return Failure(:invalid_password) if password != password_confirmation
-
-        user = User.new(attributes(:name, :password))
-
-        return Failure(:validation_error) unless user.save
-
-        Success { { user: user } }
-      end
-    end
-
-    class Fetch < Micro::Case::Strict
-      attribute :user_id
-
-      def call!
-        user = User.find_by_id(user_id)
-
-        return Success { { user: user } } if user
-
-        Failure(:user_not_found)
-      end
-    end
-
-    class CheckPassword < Micro::Case::Strict
-      attributes :user, :password
-
-      def call!
-        return Failure(:user_must_be_persisted) if user.new_record?
-        return Failure(:wrong_password) if user.wrong_password?(password)
-
-        return Success { attributes(:user) }
-      end
-    end
-
-    Authenticate = Fetch >> CheckPassword
-  end
-
-  module Todos
-    class Create < Micro::Case::Strict
-      attributes :user, :description
-
-      def call!
-        todo = Todo.new(user_id: user.id, description: description)
-
-        return Failure(:validation_error) unless todo.save
-
-        Success { { todo: todo } }
-      end
-    end
-
-    class FetchByUser < Micro::Case::Strict
-      attributes :user, :todo_id
-
-      def call!
-        todo = Todo.find_by_id_and_user_id(todo_id, user.id)
-
-        return Success { { todo: todo } } if todo
-
-        Failure(:todo_not_found)
-      end
-    end
-
-    class SetToDone < Micro::Case::Strict
-      attribute :todo
-
-      def call!
-        if todo.pending?
-          todo.done = true
-          todo.save
-        end
-
-        return Success { attributes(:todo) }
-      end
-    end
-  end
-
-  module UserTodos
-    Create = Micro::Case::Flow([Users::Authenticate, Todos::Create])
-
-    class MarkAsDone < Micro::Case
-      flow Users::Authenticate,
-           Todos::FetchByUser,
-           Todos::SetToDone
-    end
-  end
-
   def setup
     [User, Todo].each(&:delete_all)
   end
@@ -201,7 +20,7 @@ class Micro::Case::Flow::ReducerTest < Minitest::Test
     refute_predicate(user, :new_record?)
 
     todo_created =
-      UserTodos::Create.call(user_id: user.id, password: user_password, description: 'Buy milk')
+      UserTodoList::AddItem.call(user_id: user.id, password: user_password, description: 'Buy milk')
 
     assert_success_result(todo_created)
 
@@ -213,7 +32,7 @@ class Micro::Case::Flow::ReducerTest < Minitest::Test
     assert_equal(user.id, todo.user_id)
 
     result =
-      UserTodos::MarkAsDone.call(user_id: user.id, password: user_password, todo_id: todo.id)
+      UserTodoList::MarkItemAsDone.call(user_id: user.id, password: user_password, todo_id: todo.id)
 
     todo_updated = result.value[:todo]
 
@@ -235,12 +54,12 @@ class Micro::Case::Flow::ReducerTest < Minitest::Test
     user = user_created.value[:user]
 
     result1 =
-      UserTodos::Create.call(user_id: user.id, password: '', description: 'Buy beer')
+      UserTodoList::AddItem.call(user_id: user.id, password: '', description: 'Buy beer')
 
     assert_failure_result(result1, type: :wrong_password)
 
     result2 =
-      UserTodos::Create.call(user_id: user.id, password: user_password, description: '')
+      UserTodoList::AddItem.call(user_id: user.id, password: user_password, description: '')
 
     assert_failure_result(result2, type: :validation_error)
     assert_instance_of(Todos::Create, result2.use_case)
