+++ test/support/todoing/app/models/todos/create.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Todos
+  class Create < Micro::Case::Strict
+    attributes :user, :description
+
+    def call!
+      todo = Todo.new(user_id: user.id, description: description)
+
+      return Failure(:validation_error) unless todo.save
+
+      Success { { todo: todo } }
+    end
+  end
+end
