+++ test/support/todoing/app/models/user_todo_list/add_item.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module UserTodoList
+  AddItem = Micro::Case::Flow([
+    Users::Authenticate,
+    Todos::Create
+  ])
+end
