+++ test/support/todoing/app/models/users/authenticate.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Users
+  Authenticate = Find >> ValidatePassword
+end
